Improve Docker image tagging
We want every Git tag with a 'v' prefix to become a Docker build,
regardless of how many decimal places are actually in it.

Additionally, we like long SHAs in our SHA-based tags, we want the
leading v to be preserved in the version-based tags, and it's helpful
to provide major and minor version tags as well. (Then you can depend on
say, aphex-php7.4:v9 and get the minor and patch updates you expect.)